Top_n_application/requirements.txt
ffmpeg-python==0.2.0
Flask==3.0.2
huggingface-hub==0.20.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.15.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
lazy_loader==0.3
librosa==0.10.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
multidict==6.0.4
multiprocess==0.70.15
numpy==1.26.3
pandas==2.2.0
python==3.10.12
pytube==15.0.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
systemd-python==234
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
validators==0.22.0
youtube-dl==2021.12.17
yt-dlp==2023.12.30

